requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
WTForms==2.3.3
python-dotenv==0.19.0
Werkzeug==2.0.1 